% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d-instance.R
\name{generate_2d_instance}
\alias{generate_2d_instance}
\title{Generate 2d instance}
\usage{
generate_2d_instance(no_of_points = 50, interval = c(min = -10, max = 10))
}
\description{
Generate 2d instance
}
